refactor: move mise local package logic to step builder
