+++ core/generate/context.go
@@ -45,7 +45,6 @@ type GenerateContext struct {
 	Metadata        *Metadata
 	Resolver        *resolver.Resolver
 	MiseStepBuilder *MiseStepBuilder
-	miseInstance    *mise.Mise
 
 	Logger *logger.Logger
 }
@@ -108,17 +107,6 @@ func (c *GenerateContext) GetMiseStepBuilder() *MiseStepBuilder {
 	return c.MiseStepBuilder
 }
 
-func (c *GenerateContext) getMiseInstance() (*mise.Mise, error) {
-	if c.miseInstance == nil {
-		var err error
-		c.miseInstance, err = mise.New(mise.InstallDir)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return c.miseInstance, nil
-}
-
 func (c *GenerateContext) EnterSubContext(subContext string) *GenerateContext {
 	c.SubContexts = append(c.SubContexts, subContext)
 	return c
@@ -148,16 +136,6 @@ func (c *GenerateContext) GetStepByName(name string) *StepBuilder {
 
 func (c *GenerateContext) ResolvePackages() (map[string]*resolver.ResolvedPackage, error) {
 	return c.Resolver.ResolvePackages()
-}
-
-// GetMisePackageVersions retrieves package versions from mise tool-versions files
-func (c *GenerateContext) GetMisePackageVersions() (map[string]*mise.MisePackageInfo, error) {
-	miseInstance, err := c.getMiseInstance()
-	if err != nil {
-		return nil, err
-	}
-
-	return miseInstance.GetPackageVersions(c)
 }
 
 // Generate a build plan from the context
