+++ core/mise/mise.go
@@ -1,8 +1,8 @@
+// helper utilities to run the mise tool
 package mise
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"os"
 	"os/exec"
@@ -12,7 +12,6 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/alexflint/go-filemutex"
 	"github.com/charmbracelet/log"
-	"github.com/railwayapp/railpack/core/logger"
 	"github.com/railwayapp/railpack/internal/utils"
 )
 
@@ -122,53 +121,9 @@ func (m *Mise) GetAllVersions(pkg, version string) ([]string, error) {
 	return versions, nil
 }
 
-// gets all package versions from mise that are defined in the app directory environment
-// this can include additional packages defined outside the app directory, but we filter those out
-func (m *Mise) GetPackageVersions(ctx MiseAppContext) (map[string]*MisePackageInfo, error) {
-	appDir := ctx.GetAppSource()
-	output, err := m.runCmd("--cd", appDir, "list", "--current", "--json")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get package versions: %w", err)
-	}
-
-	var listOutput MisePackageListOutput
-	if err := json.Unmarshal([]byte(output), &listOutput); err != nil {
-		return nil, fmt.Errorf("failed to parse mise list output: %w", err)
-	}
-
-	packages := make(map[string]*MisePackageInfo)
-
-	for toolName, tools := range listOutput {
-		var appDirTools []MiseListTool
-		for _, tool := range tools {
-			// Only include tools that are sourced from within the app directory
-			if strings.HasPrefix(tool.Source.Path, appDir) {
-				appDirTools = append(appDirTools, tool)
-			}
-		}
-
-		if len(appDirTools) > 1 {
-			versions := make([]string, len(appDirTools))
-			for i, tool := range appDirTools {
-				versions[i] = tool.Version
-			}
-
-			// this is possible, although in practice it should be extremely rare
-			ctx.GetLogger().LogWarn("Multiple versions of tool '%s' found: %v. Using the first one: %s",
-				toolName, versions, versions[0])
-		}
-
-		if len(appDirTools) > 0 {
-			firstTool := appDirTools[0]
-			packages[toolName] = &MisePackageInfo{
-				Version: firstTool.Version,
-				// include the source so we can surface this to the user so they understand where the package version came from
-				Source: firstTool.Source.Type,
-			}
-		}
-	}
-
-	return packages, nil
+// GetCurrentList returns the JSON output of 'mise list --current --json' for a specific app directory
+func (m *Mise) GetCurrentList(appDir string) (string, error) {
+	return m.runCmd("--cd", appDir, "list", "--current", "--json")
 }
 
 // runCmd runs a mise command with the given arguments
@@ -214,36 +169,10 @@ type MiseConfig struct {
 	Tools map[string]MisePackage `toml:"tools"`
 }
 
-// MiseListSource represents the source of a mise tool installation
-type MiseListSource struct {
-	Type string `json:"type"`
-	Path string `json:"path"`
-}
-
-// represents a tool in the mise list output
-type MiseListTool struct {
-	Version          string         `json:"version"`
-	RequestedVersion string         `json:"requested_version"`
-	InstallPath      string         `json:"install_path"`
-	Source           MiseListSource `json:"source"`
-	Installed        bool           `json:"installed"`
-	// --current ensures Active=true for all entries
-	Active bool `json:"active"`
-}
-
-// full output of `mise list --current --json`
-type MisePackageListOutput map[string][]MiseListTool
-
 // represents a app-local mise package
 type MisePackageInfo struct {
 	Version string
 	Source  string
-}
-
-// a separate interface instead of *generate.GenerateContext directly to avoid import cycling
-type MiseAppContext interface {
-	GetAppSource() string
-	GetLogger() *logger.Logger
 }
 
 func GenerateMiseToml(packages map[string]string) (string, error) {
